I replace the previous by the code including the resource monitoring file
