+++ scripts/merge_by_group.py
@@ -1,8 +1,8 @@
 from pyspark.sql import functions as F
 
 
-def cleanse_stocks(df, column, group):
-    '''Filters DataFrame for particular group by relevant dates, adds columns by Year and Week.'''
+def process_stocks(df, column, group):
+    '''Filters DataFrame for particular group by relevant dates and groups by Year and Week.'''
     print(f'\nCleansing data for stocks in {group}...')
 
     # Only select necessary columns
@@ -24,23 +24,23 @@ def cleanse_stocks(df, column, group):
         df = df.groupBy("Group", "Year", "Week").agg(F.sum(column).alias(column))
     else:
         df = df.groupBy("Group", "Year", "Week").agg(F.avg(column).alias(column))
-   
+
     # Remove columns with null values in any row
     return df.na.drop()
 
 
 def merge_by_group(stock_df, stock_column, stock_group, covid_df, write_path_final):
-    '''Groups DataFrames for given stock market / sector by Year and Week calculating the average   
+    '''Groups DataFrames for given stock market / sector by Year and Week calculating the average
     value for stock_column. Then merges with Covid data and returns the merged DataFrame.'''
 
-    # Cleanse stock market data
-    stock_df = cleanse_stocks(stock_df, stock_column, stock_group)
+    # Get stock market data
+    stock_df = process_stocks(stock_df, stock_column, stock_group)
 
     # Merge with Covid data
     df = stock_df.join(covid_df, on=["Year", "Week"], how='leftouter')
 
     # Write to CSV file
     print(f'Writing to {write_path_final} ...')
-    df.write.csv(write_path_final, header=True, mode="overwrite")     
+    df.write.csv(write_path_final, header=True, mode="overwrite")
 
     return df
