+++ scripts/modify_stocks.py
