+++ scripts/resource_monitoring.py
